Specify readiness probe for storetheindex
This is to reduce disruption in service while the indexer is
re-instantiating by reading disk which may take some time.

Builds on top of templated value here:
- https://github.com/filecoin-project/helm-charts/pull/139

Note that the PR linked above also fixes the templating of resources,
which is specified in values file. This means as part of merging this PR
the resources will now take effect, whereas before they had no impact on
the pods running.