Set calibnet imagepolicy to filecoin/all-in-one
Also add extraction to the image matching so we scan calibnet release
candidates but only worry about matching semver cores within those.
Note that this is an incomplete regex match and it's based on the
current usage pattern in this registry. For a fully correct solution
we should match against the regex supplied by https://semver.org/
but imho this is too untidy to include in the yaml manifest and
won't in practice give us much benefit. fixes #1124