Use the all-in-one lotus dockerfile for simpler/faster build
Use the default dockerfile already offered by lotus repo to build the
binaries. This, combined with heavy use of caching speeds up the lotus
binary build significantly.

Remove all references to the old dockerfiles in the repo and update
scripts as needed.

While at it, remove all redundant and unused scripts that relate to
building lotus.